+++ app/components/exhibition/exhibitionView.js
